requirements.txt
qwen_agent
dashscope
requests
selenium
Pillow
webdriver_manager
json5